股票源数据库字段更新

+++ model/src/main/java/com/ljq/stock/alert/model/entity/StockSourceEntity.java
@@ -95,7 +95,7 @@ public class StockSourceEntity extends BaseEntity {
      **/
     @TableField(value = "TRADE_NUMBER")
     @ApiModelProperty(value = "交易量", name = "tradeNumber")
-    private Integer tradeNumber;
+    private Long tradeNumber;
     /**
      * 交易金额
      **/
